Move mutator interfaces to a separate service
