+++ extensions/amp-story/0.1/amp-story-info-dialog.js
@@ -63,8 +63,8 @@ export class InfoDialog {
     /** @private @const {!Element} */
     this.parentEl_ = parentEl;
 
-    /** @private @const {!../../../src/service/resources-interface.ResourcesInterface} */
-    this.resources_ = Services.resourcesForDoc(getAmpdoc(this.win_.document));
+    /** @private @const {!../../../src/service/mutator-interface.MutatorInterface} */
+    this.mutator_ = Services.mutatorForDoc(getAmpdoc(this.win_.document));
 
     /** @private {?Element} */
     this.moreInfoLinkEl_ = null;
@@ -103,7 +103,7 @@ export class InfoDialog {
       '.i-amphtml-story-info-dialog-container'
     );
 
-    const appendPromise = this.resources_.mutateElement(this.parentEl_, () => {
+    const appendPromise = this.mutator_.mutateElement(this.parentEl_, () => {
       this.parentEl_.appendChild(root);
     });
 
@@ -148,7 +148,7 @@ export class InfoDialog {
    * @private
    */
   onInfoDialogStateUpdated_(isOpen) {
-    this.resources_.mutateElement(dev().assertElement(this.element_), () => {
+    this.mutator_.mutateElement(dev().assertElement(this.element_), () => {
       this.element_.classList.toggle(DIALOG_VISIBLE_CLASS, isOpen);
     });
   }
@@ -204,7 +204,7 @@ export class InfoDialog {
       this.element_.querySelector('.i-amphtml-story-info-heading')
     );
 
-    return this.resources_.mutateElement(headingEl, () => {
+    return this.mutator_.mutateElement(headingEl, () => {
       headingEl.textContent = label;
     });
   }
@@ -219,7 +219,7 @@ export class InfoDialog {
       this.element_.querySelector('.i-amphtml-story-info-link')
     );
 
-    return this.resources_.mutateElement(linkEl, () => {
+    return this.mutator_.mutateElement(linkEl, () => {
       linkEl.setAttribute('href', pageUrl);
 
       // Add zero-width space character (\u200B) after "." and "/" characters
@@ -242,7 +242,7 @@ export class InfoDialog {
       this.element_.querySelector('.i-amphtml-story-info-moreinfo')
     );
 
-    return this.resources_.mutateElement(this.moreInfoLinkEl_, () => {
+    return this.mutator_.mutateElement(this.moreInfoLinkEl_, () => {
       const label = this.localizationService_.getLocalizedString(
         LocalizedStringId.AMP_STORY_DOMAIN_DIALOG_HEADING_LINK
       );
