+++ extensions/amp-story/1.0/amp-story-affiliate-link.js
@@ -59,8 +59,8 @@ export class AmpStoryAffiliateLink {
     /** @private {string} */
     this.text_ = this.element_.textContent;
 
-    /** @private @const {!../../../src/service/resources-interface.ResourcesInterface} */
-    this.resources_ = Services.resourcesForDoc(getAmpdoc(this.win_.document));
+    /** @private @const {!../../../src/service/mutator-interface.MutatorInterface} */
+    this.mutator_ = Services.mutatorForDoc(getAmpdoc(this.win_.document));
   }
 
   /**
@@ -71,7 +71,7 @@ export class AmpStoryAffiliateLink {
       return;
     }
 
-    this.resources_.mutateElement(this.element_, () => {
+    this.mutator_.mutateElement(this.element_, () => {
       this.element_.textContent = '';
       this.element_.toggleAttribute('pristine', true);
       this.addPulseElement_();
