+++ extensions/amp-story/1.0/amp-story-embedded-component.js
@@ -283,8 +283,8 @@ export class AmpStoryEmbeddedComponent {
     /** @private @const {!./amp-story-store-service.AmpStoryStoreService} */
     this.storeService_ = getStoreService(this.win_);
 
-    /** @private @const {!../../../src/service/resources-interface.ResourcesInterface} */
-    this.resources_ = Services.resourcesForDoc(getAmpdoc(this.win_.document));
+    /** @private @const {!../../../src/service/mutator-interface.MutatorInterface} */
+    this.mutator_ = Services.mutatorForDoc(getAmpdoc(this.win_.document));
 
     /** @private @const {!../../../src/service/owners-interface.OwnersInterface} */
     this.owners_ = Services.ownersForDoc(getAmpdoc(this.win_.document));
@@ -424,7 +424,7 @@ export class AmpStoryEmbeddedComponent {
   toggleExpandedView_(targetToExpand) {
     if (!targetToExpand) {
       this.expandedViewOverlay_ &&
-        this.resources_.mutateElement(this.expandedViewOverlay_, () => {
+        this.mutator_.mutateElement(this.expandedViewOverlay_, () => {
           this.componentPage_.classList.toggle(
             'i-amphtml-expanded-mode',
             false
@@ -443,7 +443,7 @@ export class AmpStoryEmbeddedComponent {
     if (!this.expandedViewOverlay_) {
       this.buildAndAppendExpandedViewOverlay_();
     }
-    this.resources_.mutateElement(
+    this.mutator_.mutateElement(
       dev().assertElement(this.expandedViewOverlay_),
       () => {
         toggle(dev().assertElement(this.expandedViewOverlay_), true);
@@ -458,9 +458,8 @@ export class AmpStoryEmbeddedComponent {
    */
   buildAndAppendExpandedViewOverlay_() {
     this.expandedViewOverlay_ = buildExpandedViewOverlay(this.storyEl_);
-    this.resources_.mutateElement(
-      dev().assertElement(this.componentPage_),
-      () => this.componentPage_.appendChild(this.expandedViewOverlay_)
+    this.mutator_.mutateElement(dev().assertElement(this.componentPage_), () =>
+      this.componentPage_.appendChild(this.expandedViewOverlay_)
     );
   }
 
@@ -530,7 +529,7 @@ export class AmpStoryEmbeddedComponent {
    */
   onFocusedStateUpdate_(component) {
     if (!component) {
-      this.resources_.mutateElement(
+      this.mutator_.mutateElement(
         dev().assertElement(this.focusedStateOverlay_),
         () => {
           this.focusedStateOverlay_.classList.toggle('i-amphtml-hidden', true);
@@ -566,7 +565,7 @@ export class AmpStoryEmbeddedComponent {
       this.storyEl_.querySelector('amp-story-page[active]')
     );
 
-    this.resources_.mutateElement(
+    this.mutator_.mutateElement(
       dev().assertElement(this.focusedStateOverlay_),
       () => {
         this.focusedStateOverlay_.classList.toggle('i-amphtml-hidden', false);
@@ -618,7 +617,7 @@ export class AmpStoryEmbeddedComponent {
    * @private
    */
   onUIStateUpdate_(uiState) {
-    this.resources_.mutateElement(
+    this.mutator_.mutateElement(
       dev().assertElement(this.focusedStateOverlay_),
       () => {
         [UIType.DESKTOP_FULLBLEED, UIType.DESKTOP_PANELS].includes(uiState)
@@ -755,7 +754,7 @@ export class AmpStoryEmbeddedComponent {
       `Failed to look up embed style element with ID ${embedId}`
     );
     const embedData = embedStyleEl[AMP_EMBED_DATA];
-    this.resources_.measureMutateElement(
+    this.mutator_.measureMutateElement(
       target,
       /** measure */
       () => {
@@ -801,9 +800,9 @@ export class AmpStoryEmbeddedComponent {
    * that content around stays put.
    * @param {!Element} pageEl
    * @param {!Element} element
-   * @param {!../../../src/service/resources-interface.ResourcesInterface} resources
+   * @param {!../../../src/service/mutator-interface.MutatorInterface} mutator
    */
-  static prepareForAnimation(pageEl, element, resources) {
+  static prepareForAnimation(pageEl, element, mutator) {
     let elId = null;
 
     // When a window resize happens, we must reset the styles and prepare the
@@ -819,7 +818,7 @@ export class AmpStoryEmbeddedComponent {
     }
 
     const state = {};
-    resources.measureMutateElement(
+    mutator.measureMutateElement(
       element,
       /** measure */
       () => {
@@ -910,7 +909,7 @@ export class AmpStoryEmbeddedComponent {
       '.i-amphtml-tooltip-action-icon'
     );
 
-    this.resources_.mutateElement(dev().assertElement(actionIcon), () => {
+    this.mutator_.mutateElement(dev().assertElement(actionIcon), () => {
       actionIcon.classList.toggle(embedConfig.actionIcon, true);
     });
   }
@@ -941,7 +940,7 @@ export class AmpStoryEmbeddedComponent {
 
     // Publisher specified a valid icon url.
     if (iconUrl) {
-      this.resources_.mutateElement(
+      this.mutator_.mutateElement(
         dev().assertElement(tooltipCustomIcon),
         () => {
           setImportantStyles(dev().assertElement(tooltipCustomIcon), {
@@ -953,12 +952,9 @@ export class AmpStoryEmbeddedComponent {
     }
 
     // No icon src specified by publisher. Use default icon found in the config.
-    this.resources_.mutateElement(
-      dev().assertElement(tooltipCustomIcon),
-      () => {
-        tooltipCustomIcon.classList.add(embedConfig.customIconClassName);
-      }
-    );
+    this.mutator_.mutateElement(dev().assertElement(tooltipCustomIcon), () => {
+      tooltipCustomIcon.classList.add(embedConfig.customIconClassName);
+    });
   }
 
   /**
@@ -996,7 +992,7 @@ export class AmpStoryEmbeddedComponent {
   positionTooltip_(component) {
     const state = {arrowOnTop: false};
 
-    this.resources_.measureMutateElement(
+    this.mutator_.measureMutateElement(
       this.storyEl_,
       /** measure */
       () => {
@@ -1096,7 +1092,7 @@ export class AmpStoryEmbeddedComponent {
    * @private
    */
   clearTooltip_() {
-    this.resources_.mutateElement(dev().assertElement(this.tooltip_), () => {
+    this.mutator_.mutateElement(dev().assertElement(this.tooltip_), () => {
       const actionIcon = this.tooltip_.querySelector(
         '.i-amphtml-tooltip-action-icon'
       );
