+++ extensions/amp-story/1.0/test/test-amp-story-access.js
@@ -47,7 +47,7 @@ describes.realWin('amp-story-access', {amp: true}, env => {
 
     storyEl = win.document.createElement('amp-story');
     const storyAccessEl = win.document.createElement('amp-story-access');
-    storyAccessEl.getResources = () => win.__AMP_SERVICES.resources.obj;
+    storyAccessEl.getMutator = () => win.__AMP_SERVICES.mutator.obj;
 
     storyEl.appendChild(storyAccessEl);
