+++ extensions/amp-story/1.0/test/test-amp-story-consent.js
@@ -83,7 +83,7 @@ describes.realWin('amp-story-consent', {amp: true}, env => {
     setConfig(defaultConfig);
 
     storyConsentEl = win.document.createElement('amp-story-consent');
-    storyConsentEl.getResources = () => win.__AMP_SERVICES.resources.obj;
+    storyConsentEl.getMutator = () => win.__AMP_SERVICES.mutator.obj;
     storyConsentEl.appendChild(storyConsentConfigEl);
 
     storyEl.appendChild(consentEl);
