+++ extensions/amp-story/1.0/test/test-amp-story-embedded-component.js
@@ -44,7 +44,7 @@ describes.realWin('amp-story-embedded-component', {amp: true}, env => {
     addAttributesToElement(clickableEl, {'href': 'https://google.com'});
 
     // Making sure resource tasks run synchronously.
-    sandbox.stub(Services, 'resourcesForDoc').returns({
+    sandbox.stub(Services, 'mutatorForDoc').returns({
       mutateElement: (element, callback) => {
         callback();
         return Promise.resolve();
