+++ src/service/resources-interface.js
@@ -14,16 +14,37 @@
  * limitations under the License.
  */
 
-import {MutatorInterface} from './mutator-interface';
-
 /** @const {string} */
 export const READY_SCAN_SIGNAL = 'ready-scan';
+
+/**
+ * The internal structure of a ChangeHeightRequest.
+ * @typedef {{
+ *   newMargins: !../layout-rect.LayoutMarginsChangeDef,
+ *   currentMargins: !../layout-rect.LayoutMarginsDef
+ * }}
+ */
+export let MarginChangeDef;
+
+/**
+ * The internal structure of a ChangeHeightRequest.
+ * @typedef {{
+ *   resource: !./resource.Resource,
+ *   newHeight: (number|undefined),
+ *   newWidth: (number|undefined),
+ *   marginChange: (!MarginChangeDef|undefined),
+ *   event: (?Event|undefined),
+ *   force: boolean,
+ *   callback: (function(boolean)|undefined),
+ * }}
+ */
+export let ChangeSizeRequestDef;
 
 /* eslint-disable no-unused-vars */
 /**
  * @interface
  */
-export class ResourcesInterface extends MutatorInterface {
+export class ResourcesInterface {
   /**
    * Returns a list of resources.
    * @return {!Array<!./resource.Resource>}
@@ -107,6 +128,18 @@ export class ResourcesInterface extends MutatorInterface {
   schedulePass(opt_delay, opt_relayoutAll) {}
 
   /**
+   * Enqueue, or update if already exists, a mutation task for a resource.
+   * @param {./resource.Resource} resource
+   * @param {ChangeSizeRequestDef} newRequest
+   */
+  updateOrEnqueueMutateTask(resource, newRequest) {}
+
+  /**
+   * Schedules the work pass at the latest with the specified delay.
+   */
+  schedulePassVsync() {}
+
+  /**
    * Registers a callback to be called when the next pass happens.
    * @param {function()} callback
    */
@@ -122,6 +155,16 @@ export class ResourcesInterface extends MutatorInterface {
    * May never be called in Shadow Mode.
    */
   ampInitComplete() {}
+
+  /**
+   * @param {number} relayoutTop
+   */
+  setRelayoutTop(relayoutTop) {}
+
+  /**
+   * Flag that the height could have been changed.
+   */
+  maybeHeightChanged() {}
 
   /**
    * Updates the priority of the resource. If there are tasks currently
