+++ src/service/standard-actions-impl.js
@@ -78,8 +78,8 @@ export class StandardActions {
       ? opt_win.document.documentElement
       : ampdoc.getHeadNode();
 
-    /** @const @private {!./resources-interface.ResourcesInterface} */
-    this.resources_ = Services.resourcesForDoc(ampdoc);
+    /** @const @private {!./mutator-interface.MutatorInterface} */
+    this.mutator_ = Services.mutatorForDoc(ampdoc);
 
     /** @const @private {!./viewport/viewport-interface.ViewportInterface} */
     this.viewport_ = Services.viewportForDoc(ampdoc);
@@ -319,7 +319,7 @@ export class StandardActions {
   handleHide_(invocation) {
     const target = dev().assertElement(invocation.node);
 
-    this.resources_.mutateElement(target, () => {
+    this.mutator_.mutateElement(target, () => {
       if (target.classList.contains('i-amphtml-element')) {
         target./*OK*/ collapse();
       } else {
@@ -350,7 +350,7 @@ export class StandardActions {
       return null;
     }
 
-    this.resources_.measureElement(() => {
+    this.mutator_.measureElement(() => {
       if (
         computedStyle(ownerWindow, target).display == 'none' &&
         !isShowable(target)
@@ -370,7 +370,7 @@ export class StandardActions {
     if (autofocusElOrNull && Services.platformFor(ownerWindow).isIos()) {
       this.handleShowSync_(target, autofocusElOrNull);
     } else {
-      this.resources_.mutateElement(target, () => {
+      this.mutator_.mutateElement(target, () => {
         this.handleShowSync_(target, autofocusElOrNull);
       });
     }
@@ -425,7 +425,7 @@ export class StandardActions {
       return null;
     }
 
-    this.resources_.mutateElement(target, () => {
+    this.mutator_.mutateElement(target, () => {
       if (args['force'] !== undefined) {
         // must be boolean, won't do type conversion
         const shouldForce = user().assertBoolean(
